feat(e2ee) add ExternallyManagedKeyHandler
