+++ modules/e2ee/E2EEContext.js
@@ -23,8 +23,9 @@ const kJitsiE2EE = Symbol('kJitsiE2EE');
 export default class E2EEcontext {
     /**
      * Build a new E2EE context instance, which will be used in a given conference.
+     * @param {boolean} [options.sharedKey] - whether there is a uniques key shared amoung all participants.
      */
-    constructor() {
+    constructor({ sharedKey } = {}) {
         // Determine the URL for the worker script. Relative URLs are relative to
         // the entry point, not the script that launches the worker.
         let baseUrl = '';
@@ -44,7 +45,13 @@ export default class E2EEcontext {
         const blobUrl = window.URL.createObjectURL(workerBlob);
 
         this._worker = new Worker(blobUrl, { name: 'E2EE Worker' });
+
         this._worker.onerror = e => logger.error(e);
+
+        this._worker.postMessage({
+            operation: 'initialize',
+            sharedKey
+        });
     }
 
     /**
@@ -57,6 +64,16 @@ export default class E2EEcontext {
         this._worker.postMessage({
             operation: 'cleanup',
             participantId
+        });
+    }
+
+    /**
+     * Cleans up all state associated with all participants in the conference. This is needed when disabling e2ee.
+     *
+     */
+    cleanupAll() {
+        this._worker.postMessage({
+            operation: 'cleanupAll'
         });
     }
 
@@ -136,9 +153,9 @@ export default class E2EEcontext {
     setKey(participantId, key, keyIndex) {
         this._worker.postMessage({
             operation: 'setKey',
-            participantId,
             key,
-            keyIndex
+            keyIndex,
+            participantId
         });
     }
 }
